fix: merge weapon selection tests in one file
